Renamed and reorganized, added random level generation files
Renamed the GameObjectTree project into BackBone, for it is the backbone
of our operation.
Added a Gamespecific project for Gamespecific classes.
Added random level generation files.
#WARNING#
STILL A MAIN METHOD IN CLASS LEVEL!!!

random levels generate a list of keys and a dictionary of tiles. The
list with keys can be used to get a key to get a Tile from the
dictionary if you do not know a key yet.

The Code is still a bit messy and requires a cleanup in the near future.