Level Counter, Random Levels chased, Random Levels basic efficiëncy upped
Random Levels:

Edited: Level.cs / RandomLevel.cs / PlayingState.cs

Added a parameter for RandomLevel for when being "chased". If so then
the sidepaths will be significally smaller and straigth.

Added a couple of variables to check which level it is that will be
made.

The efficiëncy of RandomLevel  is increased by doing a couple of things
in a more efficient way.